Many changes related to the change to being a backend JSON authentication API
Restructuring to be a REST API for a single-page HTML application or back-end for an app. Rather than responses being rendered into HTML using templates, assumption is that the client can handle JSON requests and responses for everything but login (using HTTP Basic) and the email address verification link (which has to be handled in a browser).

Account related endpoints are in /auth handled by accounts.js, the server API we're authenticating (which does nothing more than echo a JSON POST body)  is in api.js.

Work in progress.